pyproject.toml
[project]
name = "tracelet"
version = "0.0.1"
description = "Its a tracking app that tracks packages cuz dpd tracking sucks"
requires-python = ">=3.14"
dependencies = [
    "alembic==1.18.3",
    "fastapi==0.128.1",
    "psycopg2-binary==2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv==1.2.1",
    "sqlalchemy==2.0.46",
    "uvicorn>=0.40.0",
]
